Re-enable benchmarks
Due to changes in how `Buffer` is created, I decided to move benchmarks
side by side to the code so it's possible to reuse `Test*` traits to
easily measure. Moreover there is a crate less to be installed.

Ergonomics are not the best since all ignored tests are being included
in the output of `cargo bench` and I'm thinking of a way to improve
this.

Interestingly results shows a big issue in `BufferView.add_selection_above`.
Same measurements are fine when performing a similar test directly on
the `Buffer`.