Get `replica_id` from the client
The `replica_id` is now passed by the remote client to the application
when connecting to a Workspace. This ensures that all operations
produced by a user will be linked to him.

During boot in the browser, a `replica_id` retrieved or created and
saved in the localStorage. In the future, a "login with GitHub" button
should be shown to retrieve and use the `user_id` as the replica one.

With this commit `UserId` has been removed and all references to a
user will be dictated by a `ReplicaId` instead.